requirements.txt
beautifulsoup4==4.12.2
numpy==1.23.5
openai==1.2.2
pandas==2.0.1
requests==2.31.0
scipy==1.11.3
tiktoken==0.5.1
faiss-cpu==1.7.4
